Revert "Merge branch 'refactor_cnn' of https://github.com/maxwellsoh/emgBenchmarking into refactor_cnn"
This reverts commit e2f2ece2e3dfb730d8b56759925a5400a5ce3249, reversing
changes made to 061e165819129589e54ea28017c149325386047e.
